+++ compiler-rt/lib/gwp_asan/guarded_pool_allocator.h
@@ -13,6 +13,7 @@
 #include "gwp_asan/definitions.h"
 #include "gwp_asan/mutex.h"
 #include "gwp_asan/options.h"
+#include "gwp_asan/platform_specific/guarded_pool_allocator_tls.h"
 #include "gwp_asan/stack_trace_compressor.h"
 
 #include <stddef.h>
@@ -77,12 +78,12 @@ public:
     // class must be valid when zero-initialised, and we wish to sample as
     // infrequently as possible when this is the case, hence we underflow to
     // UINT32_MAX.
-    if (GWP_ASAN_UNLIKELY(ThreadLocals.NextSampleCounter == 0))
-      ThreadLocals.NextSampleCounter =
+    if (GWP_ASAN_UNLIKELY(getThreadLocals()->NextSampleCounter == 0))
+      getThreadLocals()->NextSampleCounter =
           ((getRandomUnsigned32() % (AdjustedSampleRatePlusOne - 1)) + 1) &
           ThreadLocalPackedVariables::NextSampleCounterMask;
 
-    return GWP_ASAN_UNLIKELY(--ThreadLocals.NextSampleCounter == 0);
+    return GWP_ASAN_UNLIKELY(--getThreadLocals()->NextSampleCounter == 0);
   }
 
   // Returns whether the provided pointer is a current sampled allocation that
@@ -206,37 +207,10 @@ private:
   // the sample rate.
   uint32_t AdjustedSampleRatePlusOne = 0;
 
-  // Pack the thread local variables into a struct to ensure that they're in
-  // the same cache line for performance reasons. These are the most touched
-  // variables in GWP-ASan.
-  struct alignas(8) ThreadLocalPackedVariables {
-    constexpr ThreadLocalPackedVariables()
-        : RandomState(0xff82eb50), NextSampleCounter(0), RecursiveGuard(false) {
-    }
-    // Initialised to a magic constant so that an uninitialised GWP-ASan won't
-    // regenerate its sample counter for as long as possible. The xorshift32()
-    // algorithm used below results in getRandomUnsigned32(0xff82eb50) ==
-    // 0xfffffea4.
-    uint32_t RandomState;
-    // Thread-local decrementing counter that indicates that a given allocation
-    // should be sampled when it reaches zero.
-    uint32_t NextSampleCounter : 31;
-    // The mask is needed to silence conversion errors.
-    static const uint32_t NextSampleCounterMask = (1U << 31) - 1;
-    // Guard against recursivity. Unwinders often contain complex behaviour that
-    // may not be safe for the allocator (i.e. the unwinder calls dlopen(),
-    // which calls malloc()). When recursive behaviour is detected, we will
-    // automatically fall back to the supporting allocator to supply the
-    // allocation.
-    bool RecursiveGuard : 1;
-  };
-  static_assert(sizeof(ThreadLocalPackedVariables) == sizeof(uint64_t),
-                "thread local data does not fit in a uint64_t");
-
   class ScopedRecursiveGuard {
   public:
-    ScopedRecursiveGuard() { ThreadLocals.RecursiveGuard = true; }
-    ~ScopedRecursiveGuard() { ThreadLocals.RecursiveGuard = false; }
+    ScopedRecursiveGuard() { getThreadLocals()->RecursiveGuard = true; }
+    ~ScopedRecursiveGuard() { getThreadLocals()->RecursiveGuard = false; }
   };
 
   // Initialise the PRNG, platform-specific.
@@ -245,8 +219,6 @@ private:
   // xorshift (32-bit output), extremely fast PRNG that uses arithmetic
   // operations only. Seeded using platform-specific mechanisms by initPRNG().
   uint32_t getRandomUnsigned32();
-
-  static GWP_ASAN_TLS_INITIAL_EXEC ThreadLocalPackedVariables ThreadLocals;
 };
 } // namespace gwp_asan
