+++ compiler-rt/lib/gwp_asan/utilities.cpp
@@ -37,7 +37,7 @@ static size_t alignPowerOfTwo(size_t RealAllocationSize) {
 #ifdef __BIONIC__
 static constexpr AlignmentStrategy PlatformDefaultAlignment =
     AlignmentStrategy::BIONIC;
-#else // __BIONIC__
+#else  // __BIONIC__
 static constexpr AlignmentStrategy PlatformDefaultAlignment =
     AlignmentStrategy::POWER_OF_TWO;
 #endif // __BIONIC__
